+++ polis_recognizer/__init__.py
@@ -59,7 +59,6 @@ from .pdf_extraction_router import (
 from .pdfplumber_ingestion import PdfPlumberIngestionService
 from .policy_ingestion import ExtractedTextResult, PolicyIngestionService
 
-
 __version__ = "0.3.4"